Refactor poll shared logic
