+++ docs/development/debugging/gotchas/persession-read-in-mapped-computed.md
@@ -84,5 +84,5 @@ perSession doesn't.
 **Known-good references:**
 `lot-watch/main.tsx` (`sightingRows`, spot picker) and
 `parking-coordinator/main.tsx` (`adminPeopleData`/`adminSpotsData`) both bake the
-flag into the producing computed. `cozy-poll/main.tsx` maps `options`
+flag into the producing computed. `poll/cozy/main.tsx` maps `options`
 directly and needs no change.
