+++ packages/patterns/poll/shared/voting.tsx
@@ -1,0 +1,108 @@
+import { handler } from "commonfabric";
+
+import { trimmedName } from "./constants.tsx";
+import {
+  type CastVoteEvent,
+  type ClearVoteEvent,
+  type NameCell,
+  type ResetVotesEvent,
+  type User,
+  type Vote,
+  type VoteColor,
+  type VotesCell,
+} from "./types.tsx";
+
+export const castVote = handler<CastVoteEvent, {
+  votes: VotesCell;
+  myName: NameCell;
+}>(({ optionId, voteType }, { votes, myName }) => {
+  const me = trimmedName(myName.get());
+  if (!me) return;
+  const current = votes.get();
+  const existingIdx = current.findIndex(
+    (v) => v.voterName === me && v.optionId === optionId,
+  );
+  if (existingIdx >= 0) {
+    const existing = current[existingIdx];
+    if (existing.voteType === voteType) {
+      votes.remove(existing);
+      return;
+    }
+    votes.key(existingIdx).key("voteType").set(voteType);
+    return;
+  }
+  votes.push({ voterName: me, optionId, voteType });
+});
+
+export const resetVotes = handler<ResetVotesEvent, {
+  votes: VotesCell;
+  myName: NameCell;
+  adminName: NameCell;
+}>((_, { votes, myName, adminName }) => {
+  const me = trimmedName(myName.get());
+  const admin = trimmedName(adminName.get());
+  if (!me || me !== admin) return;
+  votes.set([]);
+});
+
+export const clearMyVote = handler<ClearVoteEvent, {
+  votes: VotesCell;
+  myName: NameCell;
+}>(({ optionId }, { votes, myName }) => {
+  const me = trimmedName(myName.get());
+  if (!me) return;
+  votes.set(
+    votes.get().filter(
+      (v) => !(v.voterName === me && v.optionId === optionId),
+    ),
+  );
+});
+
+export interface TallyableOption {
+  id: string;
+}
+
+export interface OptionTally<TOption extends TallyableOption> {
+  option: TOption;
+  green: number;
+  yellow: number;
+  red: number;
+  voters: Array<{ name: string; voteType: VoteColor; color: string }>;
+}
+
+export const tallyOptions = <TOption extends TallyableOption>(
+  options: readonly TOption[],
+  votes: readonly Vote[],
+  users: readonly User[],
+): OptionTally<TOption>[] => {
+  const colorByName = new Map(users.map((u) => [u.name, u.color]));
+  const tallies = options.map((option): OptionTally<TOption> => {
+    const optionVotes = votes.filter((v) => v.optionId === option.id);
+    return {
+      option,
+      green: optionVotes.filter((v) => v.voteType === "green").length,
+      yellow: optionVotes.filter((v) => v.voteType === "yellow").length,
+      red: optionVotes.filter((v) => v.voteType === "red").length,
+      voters: optionVotes.map((v) => ({
+        name: v.voterName,
+        voteType: v.voteType,
+        color: colorByName.get(v.voterName) ?? "#888",
+      })),
+    };
+  });
+  return [...tallies].sort((a, b) => {
+    if (a.red !== b.red) return a.red - b.red;
+    return b.green - a.green;
+  });
+};
+
+export const myVoteFor = (
+  votes: readonly Vote[],
+  me: string,
+  optionId: string,
+): VoteColor | undefined => {
+  if (!me) return undefined;
+  return votes.find(
+    (v) => v.voterName === me && v.optionId === optionId,
+  )?.voteType;
+};
