+++ packages/ts-transformers/README.md
@@ -95,11 +95,11 @@ env FIXTURE=map-single-capture deno task test
 To see transformed output:
 
 ```bash
-deno task cf check --show-transformed [--no-run] packages/patterns/lunch-poll/main.tsx
+deno task cf check --show-transformed [--no-run] packages/patterns/poll/lunch/main.tsx
 ```
 
 Run it from the repository root, targeting a pattern that lives in the workspace
-(the path is repo-root-relative, e.g. `packages/patterns/lunch-poll/main.tsx`):
+(the path is repo-root-relative, e.g. `packages/patterns/poll/lunch/main.tsx`):
 `deno task cf` resolves the pattern against the workspace to execute the
 transform. Add `--no-run` to emit the transformed output without running the
 pattern.
